Make sure new serial data is always appended
If the user selected text and then new serial data was received,
this text would get inserted at the start of the selected text,
NOT at the end of the buffer. We fix this by creating a custom
text mark "end" that has left-gravity, so we can keep inserting
at that position and the mark will always be at the end of the
buffer.